requirements.txt
python-dotenv
google-generativeai
langchain
pyPDF2
faiss-cpu
streamlit
